Move utility methods to super class
